Open edit card dialog correctly
