+++ src/app/components/molecules/kwartet-card-preview/edit-kwartet-card.component.html
@@ -29,3 +29,16 @@
 <ng-template #notLoaded>
   <p>The set could not be loaded.</p>
 </ng-template>
+
+
+
+<app-edit-card-dialog></app-edit-card-dialog>
+<p-dialog header="Edit card" [(visible)]="editCardDialogIsVisible" [modal]="true" [draggable]="false"
+          [resizable]="false">
+  <div>
+    <p>{{currentKwartetCard?.name}}</p>
+  </div>
+  <ng-template pTemplate="footer">
+    <p-button (click)="onClickSaveCardButton()" [disabled]=cardFormControl.invalid>Save</p-button>
+  </ng-template>
+</p-dialog>
