+++ src/app/components/molecules/kwartet-card-preview/edit-kwartet-card.component.scss
