+++ src/app/components/organisms/edit-card-dialog/edit-card-dialog.component.ts
@@ -1,7 +1,5 @@
 import { Component, Input } from '@angular/core';
 import { KwartetCard } from "../../../data/models/kwartetcard-model";
-import { FormControl, Validators } from "@angular/forms";
-import { GAME_TITLE_CHARACTER_LIMIT } from "../../../config/global-settings";
 
 @Component({
   selector: 'app-edit-card-dialog',
@@ -15,19 +13,5 @@ export class EditCardDialogComponent {
 
   @Input() currentEditedCard?: KwartetCard;
 
-  // TODO
-  cardFormControl = new FormControl('', {
-    nonNullable: true,
-    validators: [Validators.required,
-      Validators.maxLength(GAME_TITLE_CHARACTER_LIMIT)]
-  });
 
-  async onClickSaveCardButton() {
-    this.editCardDialogIsVisible = false;
-    await this.saveCard();
-  }
-
-  async saveCard() {
-    // TODO
-  }
 }
